requirements.txt
flask==2.0.1
google-generativeai==0.3.0
paddlepaddle==2.5.2
paddleocr==2.7.0
python-dotenv==0.19.0
pdf2image==1.16.3
Pillow
numpy==1.24.3